refactor: update x-sdk-operations schema and move it to a file
