+++ src/spectral/utils/spectral-validator.js
@@ -101,11 +101,21 @@ function setupStaticAssets(staticAssets, defaultSpectralRulesetURI) {
   const content = fs.readFileSync(defaultSpectralRulesetURI, 'utf8');
   staticAssets['ibm:oas'] = content;
 
+  const parentDirectory = path.parse(defaultSpectralRulesetURI).dir;
+
+  // register schemas
+  const baseKey = 'ibm/schemas';
+  const schemaDirURI = path.join(parentDirectory, 'schemas');
+  fs.readdirSync(schemaDirURI).forEach(function(jsonSchemaFile) {
+    const schemaKey = path.join(baseKey, path.parse(jsonSchemaFile).name);
+    staticAssets[schemaKey] = fs.readFileSync(
+      path.join(schemaDirURI, jsonSchemaFile),
+      'utf8'
+    );
+  });
+
   // register custom functions
-  const customFunctionsDirURI = path.join(
-    path.parse(defaultSpectralRulesetURI).dir,
-    'ibm-oas'
-  );
+  const customFunctionsDirURI = path.join(parentDirectory, 'ibm-oas');
   fs.readdirSync(customFunctionsDirURI).forEach(function(customFunctionFile) {
     const customFunctionURI = path.join(
       customFunctionsDirURI,
